feat: add heatmap tool page
